pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bachelor-thesis-sml"
version = "0.1.0"
description = "Project for benchmarking and developing model merging methods for my bachelor thesis at the ETH Statistical Machine Learning Lab (SML)"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "datasets>=4.4.1",
    "huggingface-hub[hf-xet]>=0.36.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "transformers>=4.57.1",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.9.1",
]
cuda = [
    "torch>=2.9.1",
]
xpu = [
    "pytorch-triton-xpu>=3.5.0 ; sys_platform == 'linux' or sys_platform == 'win32'",
    "torch>=2.9.1",
]

[dependency-groups]
cluster = [
    "wandb>=0.23.0",
]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
    "wandb>=0.23.0",
]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]  # Errors, pyflakes, isort
ignore = ["E501"]  # Line too long

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v"] # verbose by default; for less output use -q

# MyPy configuration
[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true

[tool.uv.sources]
torch = [
  { index = "pytorch-xpu",  extra = "xpu"  },
  { index = "pytorch-cpu",  extra = "cpu"  },
  { index = "pytorch-cu128", extra = "cuda" },
]
pytorch-triton-xpu = [
  { index = "pytorch-xpu", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-xpu"
url = "https://download.pytorch.org/whl/xpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
environments = ["sys_platform == 'win32' or sys_platform == 'linux'"]

conflicts = [
  [
    { extra = "cpu" },
    { extra = "cuda" },
  ],
  [
    { extra = "cpu" },
    { extra = "xpu" },
  ],
  [
    { extra = "cuda" },
    { extra = "xpu" },
  ],
]
